move func out of rmnpb
